+++ src/pages/Login/Login.tsx
@@ -14,7 +14,7 @@ import { authState } from '../../recoil';
 import { JWT_KEY } from '../../constants';
 import { ErrorResponse } from '../../types';
 import axios from 'axios';
-import { useSignIn } from '../../quries';
+import { useSignIn } from '../../queries';
 import { useNavigate } from 'react-router-dom';
 import { LinkText } from '../SignUp/components';
